# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI4TJU campus guide",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/scene/ack": {
      "post": {
        "tags": [
          "model-runtime"
        ],
        "summary": "Ack",
        "operationId": "ack_api_scene_ack_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneAck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneAckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "499": {
            "description": "Additional Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/speech/asr": {
      "post": {
        "tags": [
          "speech"
        ],
        "summary": "Asr",
        "operationId": "asr_api_speech_asr_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsrRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsrResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "499": {
            "description": "Additional Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiError"
      },
...
      "AsrRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "audio": {
            "$ref": "#/components/schemas/AudioPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "session_id",
          "audio"
        ],
        "title": "AsrRequest"
      },
      "AsrResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "is_final": {
            "type": "boolean",
            "title": "Is Final"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "text",
          "is_final"
        ],
        "title": "AsrResponse"
      },
      "AudioPayload": {
        "properties": {
          "encoding": {
            "type": "string",
            "const": "base64",
            "title": "Encoding"
          },
          "mime_type": {
            "type": "string",
            "const": "audio/wav",
            "title": "Mime Type"
          },
          "sample_rate_hz": {
            "type": "integer",
            "const": 16000,
            "title": "Sample Rate Hz"
          },
          "channels": {
            "type": "integer",
            "const": 1,
            "title": "Channels"
          },
          "audio_base64": {
            "type": "string",
            "maxLength": 1500000,
            "minLength": 1,
            "title": "Audio Base64"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "encoding",
          "mime_type",
          "sample_rate_hz",
          "channels",
          "audio_base64"
        ],
        "title": "AudioPayload"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ],
        "title": "ErrorDetail"
      },
...
      "SceneAck": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "execution_failed",
                  "unsupported"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "session_id",
          "action_id",
          "status"
        ],
        "title": "SceneAck"
      },
      "SceneAckResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "recorded",
              "duplicate"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "action_id",
          "status"
        ],
        "title": "SceneAckResponse"
      },
...
    }
  }
}